[fully tested code] + rf24 working along with vehicle controller

+++ Firmware/hummingbot_firmware/Hummingbot_firmware_FreeRTOS_2/source/vehicleController/vehicleController.h
@@ -15,8 +15,10 @@ typedef int16_t    speed_cm_per_s_t;
 
 // some conversion unit
 // NOTE: change to uint16_t if your processor does not have a FPU, might suffer trimming/truncation error
-typedef float   us_per_deg_t;
-typedef float   us_s_per_mm_t;
+// NOTE*: please use uint16_t, since float still takes too long to compute
+// NOTE: TODO: fixed float should be implemented in order to achieve high speed floating calc.
+typedef uint16_t   us_per_deg_t;
+typedef uint16_t   us_s_per_mm_t;
 
 typedef enum{
     VC_STATE_UNDEFINED,
@@ -58,8 +60,8 @@ bool VC_requestThrottle(speed_cm_per_s_t reqSpd);
 bool VC_doBraking(angle_deg_t reqAng);
 bool VC_powerOff_FreeWheeling(VC_channnelName_E controller);
 
+/* NOTE: PREFERRABLY: DO NOT USE RAW for main code, they are only meant for testing and calibration */ 
 bool VC_requestThrottle_raw(pulse_us_t pw_us); // will be filtered by logic
 bool VC_requestSteering_raw(pulse_us_t pw_us);
-
 bool VC_requestPWM_force_raw(VC_channnelName_E controller, pulse_us_t pw_us); // only limited by the max/min in configuration
 #endif //(VEHICLE_CONTROLLER_H_)
